requirements.txt
google-api-python-client==2.118.0
google-auth==2.28.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
pandas==2.2.1
pandas-profiling==3.1.0
PyMySQL==1.1.1
SQLAlchemy==2.0.29